+++ src/app/api/ducket/files/route.ts
@@ -1,9 +1,11 @@
-import { QUERIES } from '~/server/db/queries';
+import { API_KEY_PERMISSIONS } from '~/lib/constants';
+import { MUTATIONS, QUERIES } from '~/server/db/queries';
 
 async function validateBearerAuth(request: Request): Promise<string | undefined> {
   const bearer = request.headers.get('Authorization');
   return bearer?.split(' ')[1];
 }
+const GET_PERMISSIONS_ALLOWED = ['all', 'read'];
 
 export async function GET(request: Request) {
   try {
@@ -13,6 +15,18 @@ export async function GET(request: Request) {
     const apiKey = await validateBearerAuth(request);
     if (!apiKey) {
       return new Response('Invalid bearer auth', { status: 401 });
+    }
+    /**
+     * Validate api key permissions
+     */
+    const [apiKeyStored] = await QUERIES.getApikey({ apiKey });
+    if (!apiKeyStored) {
+      return new Response('Invalid bearer auth', { status: 402 });
+    }
+    if (
+      !apiKeyStored.permissions.some(permission => GET_PERMISSIONS_ALLOWED.includes(permission))
+    ) {
+      return new Response('Api key permissions not allowed', { status: 403 });
     }
 
     /*
@@ -26,7 +40,25 @@ export async function GET(request: Request) {
     /*
      * Get files by project id
      */
-    const files = await QUERIES.getFilesByProjectId({ projectId: project.id });
+    const [files] = await Promise.all([
+      QUERIES.getFilesByProjectId({ projectId: project.id }),
+      MUTATIONS.updateApiKeyUsage({
+        projectId: project.id,
+        apiKey: apiKey,
+      }),
+    ]);
+
+    await Promise.all(
+      files.map(file =>
+        MUTATIONS.createActivityLog({
+          projectId: project.id,
+          userId: project.ownerId,
+          fileName: file.fileName ?? '-',
+          action: API_KEY_PERMISSIONS.read,
+        })
+      )
+    );
+
     const fileUrls = files.map(file => file.fileUrl);
 
     return new Response(JSON.stringify({ files: fileUrls }), { status: 200 });